@@ -1,4 +1,4 @@
+from .__version__ import __version__
 from .extension import MKClickExtension, makeExtension
 
-__version__ = "0.0.1"
-__all__ = ["MKClickExtension", "makeExtension"]
+__all__ = ["__version__", "MKClickExtension", "makeExtension"]
